WebModelGen/requirements.txt
joblib==1.1.0
matplotlib==3.5.2
numpy==1.23.1
pandas==1.4.3
Pillow==9.2.0
scikit_learn==1.1.1
selenium==3.141.0
tabulate==0.8.10
tqdm==4.64.0
treeinterpreter==0.2.3
xgboost==2.0.2